% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_correlation.R
\name{calculate_correlation}
\alias{calculate_correlation}
\title{Calculate Correlation}
\usage{
calculate_correlation(x, y)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector.}
}
\value{
The correlation coefficient.
}
\description{
Computes the correlation between two numeric vectors.
}
