% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mean.R
\name{calculate_mean}
\alias{calculate_mean}
\title{Calculate Mean}
\usage{
calculate_mean(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The mean of the vector.
}
\description{
This function computes the mean of a numeric vector.
}
