% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_sd.R
\name{calculate_sd}
\alias{calculate_sd}
\title{Calculate Standard Deviation}
\usage{
calculate_sd(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The standard deviation of x.
}
\description{
Computes the standard deviation of a numeric vector.
}
