% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_interval.R
\name{confidence_interval}
\alias{confidence_interval}
\title{Confidence Interval}
\usage{
confidence_interval(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A confidence interval range.
}
\description{
Computes a confidence interval for a numeric vector.
}
