% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notin.R
\name{\%!in\%}
\alias{\%!in\%}
\title{Not in Operator}
\usage{
x \%!in\% y
}
\arguments{
\item{y}{}
}
\value{
Tibble
}
\description{
Returns Everything But the Vector Specified
}
\examples{
dplyr::filter(mtcars, carb \%!in\% c(1,2,3))




}
