% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_presentable.R
\name{stats_presentable}
\alias{stats_presentable}
\title{stats_presentable}
\usage{
stats_presentable(df, x, group_var1, group_var2, rounding_digits = 1)
}
\arguments{
\item{rounding_digits}{}
}
\value{
Tibble
}
\description{

}
\examples{

Use one or two grouping variables
stats_presentable(mtcars, mpg, cyl)
stats_presentable(mtcars, mpg, cyl, am)

Adjust the rounding
stats_presentable(mtcars, mpg, cyl, am, 5)
stats_presentable(mtcars, mpg, cyl, am, rounding_digits = 5)




}
