% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_summary.R
\name{stats_summary}
\alias{stats_summary}
\title{stats_summary}
\usage{
stats_summary(df, x, group_var1, group_var2, rounding_digits = 1)
}
\arguments{
\item{rounding_digits}{}
}
\value{
Tibble
}
\description{
basic summary statistics using dplyr
}
\examples{

Use one or two grouping variables
stats_summary(mtcars, mpg, cyl)
stats_summary(mtcars, mpg, cyl, am)

Adjust the rounding
stats_summary(mtcars, mpg, cyl, am, 5)
stats_summary(mtcars, mpg, cyl, am, rounding_digits = 5)




}
